regen after fix for float types in generator for 5.0 and 5.1 apis

+++ azure-devops/azure/devops/v5_0/work/models.py
@@ -13,13 +13,13 @@ class Activity(Model):
     """Activity.
 
     :param capacity_per_day:
-    :type capacity_per_day: int
+    :type capacity_per_day: float
     :param name:
     :type name: str
     """
 
     _attribute_map = {
-        'capacity_per_day': {'key': 'capacityPerDay', 'type': 'int'},
+        'capacity_per_day': {'key': 'capacityPerDay', 'type': 'float'},
         'name': {'key': 'name', 'type': 'str'}
     }
